pyproject.toml
[project]
name = "streptofile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "polars",
    "wgs_data_manager",
    "python-dotenv",
]
packages = [
    { include = "streptofile", from = "src" }
]

[tool.setuptools.package-data]
# include everything in streptofile/db
"streptofile" = ["db/*"]

[dependency-groups]
dev = ["pytest"]

[build-system]
requires = ["uv_build>=0.6.6,<0.7"]
build-backend = "uv_build"


[project.scripts]
streptofile = "streptofile.streptofile:main"

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
